V1.01 Small refact

+++ 11. Lambda and LINQ - Exercises/26. Immune System/Immune System.cs
@@ -18,6 +18,7 @@ namespace _26.Immune_System
                 var virusName = inputData;
                 double virusStrengthOriginal = (int)(inputData.Sum(x => x) / 3.0);
                 var virusStrength = virusStrengthOriginal;
+
                 if (viruses.Contains(virusName))
                 {
                     virusStrength = (virusStrength / 3.0);
@@ -32,8 +33,8 @@ namespace _26.Immune_System
 
                 if (TimeToDefeatInSec < remainingHealth)
                 {
-                    var defeatMins = (int)TimeToDefeatInSec / 60;
-                    var defeatSecs = (int)TimeToDefeatInSec % 60;
+                    var defeatMins = TimeToDefeatInSec / 60;
+                    var defeatSecs = TimeToDefeatInSec % 60;
                     Console.WriteLine($"{virusName} defeated in {defeatMins}m {defeatSecs}s.");
                     remainingHealth -= TimeToDefeatInSec;
                     Console.WriteLine($"Remaining health: {remainingHealth}");
